+++ app/src/main/java/com/example/distancetracker/ui/activities/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.distancetracker
+package com.example.distancetracker.ui.activities
 
 import android.os.Bundle
 import com.google.android.material.bottomnavigation.BottomNavigationView
@@ -7,6 +7,7 @@ import androidx.navigation.findNavController
 import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.setupActionBarWithNavController
 import androidx.navigation.ui.setupWithNavController
+import com.example.distancetracker.R
 import com.example.distancetracker.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
