+++ app/src/main/java/com/example/distancetracker/ui/fragments/data/DashboardViewModel.kt
@@ -1,13 +1,13 @@
-package com.example.distancetracker.ui.home
+package com.example.distancetracker.ui.fragments.data
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class HomeViewModel : ViewModel() {
+class DashboardViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is home Fragment"
+        value = "This is dashboard Fragment"
     }
     val text: LiveData<String> = _text
 }
