+++ app/src/main/java/com/example/distancetracker/ui/fragments/data/HomeViewModel.kt
@@ -1,13 +1,13 @@
-package com.example.distancetracker.ui.dashboard
+package com.example.distancetracker.ui.fragments.data
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class DashboardViewModel : ViewModel() {
+class HomeViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is dashboard Fragment"
+        value = "This is home Fragment"
     }
     val text: LiveData<String> = _text
 }
