+++ app/src/main/java/com/example/distancetracker/ui/fragments/data/NotificationsViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.distancetracker.ui.notifications
+package com.example.distancetracker.ui.fragments.data
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
